Complete the CRUD for pending projects. Adding Project class
