+++ ProductManagement/WebContent/WEB-INF/web.xml
@@ -12,6 +12,11 @@
 	<url-pattern>/ProductService/*</url-pattern>
   </servlet-mapping>
   
+  <servlet-mapping>
+	<servlet-name>Jersey Web Application</servlet-name>
+	<url-pattern>/ProjectService/*</url-pattern>
+  </servlet-mapping>
+  
   
   <welcome-file-list>
     <welcome-file>index.html</welcome-file>
